pyproject.toml
[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.8.0,<4"]

[project]
name = "machin"
description = "Machin-like formulae"
version = "2025.09"
readme = "README.md"
requires-python = ">=3.10,<3.15"
license = "MIT"
authors = [
    { name = "Matthew Scroggs", email = "machin@mscroggs.co.uk" },
    { name = "Katie Steckles" }
]
dependencies = [
    "arctans>=0.2.2",
    "website-build-tools>=0.1.8",
    "PyGithub",
    "pytz",
    "pytest",
    "pyyaml",
]
packages = ["machin"]

[project.optional-dependencies]
style = ["ruff", "mypy", "pydocstyle"]

[tool.ruff]
line-length = 100
indent-width = 4

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
ignore_missing_imports = true
